Added AI class to test responsiveness -- Renamed Hero classes to "KnightController" and "KnightHealth" -- Added classes "MKController" and "EnemyHealth" for Mirror Knight -- Adjusted animator controller -- Player can no longer move while blocking -- Animations now transition much quicker, improving responsiveness.
Player responsiveness is still a bit rough, but I am just about ready to begin working on the enemy's AI. Improving player controls and responsiveness will be a continuing task throughout the development.

I have also begun prototyping a combo system for attacks, as well as a different attack for running.